* Restructured calibration module and improved docs.
