+++ qiskit_experiments/calibration/analysis/__init__.py
@@ -1,0 +1,31 @@
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2021.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""
+=========================================================================
+Calibration management (:mod:`qiskit_experiments.calibration.analysis`)
+=========================================================================
+
+.. currentmodule:: qiskit_experiments.calibration.analysis
+
+Analysis classes
+======================
+.. autosummary::
+    :toctree: ../stubs/
+
+    DragCalAnalysis
+    OscillationAnalysis
+
+"""
+
+from .drag_analysis import DragCalAnalysis
+from .oscillation_analysis import OscillationAnalysis
