+++ qiskit_experiments/calibration/drag.py
@@ -24,8 +24,8 @@ from qiskit.providers.options import Options
 
 from qiskit_experiments.base_experiment import BaseExperiment
 from qiskit_experiments.data_processing.processor_library import get_to_signal_processor
-from qiskit_experiments.calibration.exceptions import CalibrationError
-from qiskit_experiments.calibration.experiments.drag_analysis import DragCalAnalysis
+from qiskit_experiments.calibration.management.exceptions import CalibrationError
+from qiskit_experiments.calibration.analysis.drag_analysis import DragCalAnalysis
 
 
 class DragCal(BaseExperiment):
