+++ qiskit_experiments/calibration/management/calibrations.py
@@ -35,9 +35,9 @@ from qiskit.pulse import (
 )
 from qiskit.pulse.channels import PulseChannel
 from qiskit.circuit import Parameter, ParameterExpression
-from qiskit_experiments.calibration.exceptions import CalibrationError
-from qiskit_experiments.calibration.parameter_value import ParameterValue
-from qiskit_experiments.calibration.calibration_key_types import (
+from qiskit_experiments.calibration.management.exceptions import CalibrationError
+from qiskit_experiments.calibration.management.parameter_value import ParameterValue
+from qiskit_experiments.calibration.management.calibration_key_types import (
     ParameterKey,
     ParameterValueType,
     ScheduleKey,
