+++ qiskit_experiments/calibration/management/update_library.py
@@ -21,11 +21,11 @@ from qiskit.circuit import Parameter
 from qiskit.pulse import ScheduleBlock
 
 from qiskit_experiments.experiment_data import ExperimentData
-from qiskit_experiments.calibration.backend_calibrations import BackendCalibrations
-from qiskit_experiments.calibration.calibrations import Calibrations
-from qiskit_experiments.calibration.parameter_value import ParameterValue
-from qiskit_experiments.calibration.exceptions import CalibrationError
-from qiskit_experiments.calibration.calibration_key_types import ParameterValueType
+from qiskit_experiments.calibration.management.backend_calibrations import BackendCalibrations
+from qiskit_experiments.calibration.management.calibrations import Calibrations
+from qiskit_experiments.calibration.management.parameter_value import ParameterValue
+from qiskit_experiments.calibration.management.exceptions import CalibrationError
+from qiskit_experiments.calibration.management.calibration_key_types import ParameterValueType
 
 
 class BaseUpdater(ABC):
@@ -185,7 +185,7 @@ class Amplitude(BaseUpdater):
         Raises:
             CalibrationError: If the experiment is not of the supported type.
         """
-        from qiskit_experiments.calibration.experiments.rabi import Rabi
+        from qiskit_experiments.calibration.rabi import Rabi
 
         if angles_schedules is None:
             angles_schedules = [(np.pi, "amp", "xp")]
