+++ test/calibration/experiments/test_drag.py
@@ -21,8 +21,8 @@ import qiskit.pulse as pulse
 from qiskit.qobj.utils import MeasLevel
 from qiskit import transpile
 
-from qiskit_experiments.calibration.exceptions import CalibrationError
-from qiskit_experiments.calibration.experiments.drag import DragCal
+from qiskit_experiments.calibration.management.exceptions import CalibrationError
+from qiskit_experiments.calibration.drag import DragCal
 from qiskit_experiments.test.mock_iq_backend import DragBackend
