+++ test/calibration/test_backend_calibrations.py
@@ -14,7 +14,7 @@
 
 from qiskit.test import QiskitTestCase
 from qiskit.test.mock import FakeArmonk
-from qiskit_experiments.calibration.backend_calibrations import BackendCalibrations
+from qiskit_experiments.calibration.management.backend_calibrations import BackendCalibrations
 
 
 class TestBackendCalibrations(QiskitTestCase):
